requirements-dev.txt
# Runtime requirements
--requirement requirements.txt

# Testing
pytest==2.5.1
py==1.4.19
mock==1.0.1

# Linting
flake8==2.1.0
mccabe==0.2.1
pep8==1.4.6
pyflakes==0.7.3

# Documentation
Sphinx==1.2
docutils==0.11
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6

# Miscellaneous
Paver==1.2.1
colorama==0.2.7
ruffus
pyyaml
setuptools
six>=1.3
termcolor
bunch==1.0.1
python-dateutil
matplotlib
